+++ lib/ui/gies/CheckListPanel.dart
@@ -14,8 +14,12 @@
  * limitations under the License.
  */
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/ui/gies/CheckListContentWidget.dart';
+import 'package:illinois/utils/AppUtils.dart';
+import 'package:rokwire_plugin/service/auth2.dart';
+import 'package:rokwire_plugin/service/connectivity.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 
@@ -23,6 +27,34 @@ class CheckListPanel extends StatefulWidget {
   final String contentKey;
 
   const CheckListPanel({Key? key, required this.contentKey}) : super(key: key);
+
+  static void present(BuildContext context, { required String contentKey }) {
+    if (Connectivity().isOffline) {
+      AppAlert.showOfflineMessage(context, _offlineMessage(contentKey));
+    }
+    else if (!Auth2().isOidcLoggedIn) {
+      AppAlert.showMessage(context, _loggedOutMessage(contentKey));
+    }
+    else {
+      Navigator.push(context, CupertinoPageRoute(builder: (context) => CheckListPanel(contentKey: contentKey,)));
+    }
+  }
+
+  static String? _offlineMessage(String contentKey) {
+    if (contentKey == "gies") {
+      return Localization().getStringEx('widget.checklist.gies.offline', 'iDegrees New Student Checklist not available while offline.');
+    } else if (contentKey == "uiuc_student") {
+      return Localization().getStringEx('widget.checklist.uiuc.offline', 'New Student Checklist not available while offline.');
+    }
+  }
+
+  static String? _loggedOutMessage(String contentKey) {
+    if (contentKey == "gies") {
+      return Localization().getStringEx('widget.checklist.gies.logged_out', 'You need to be logged in to access iDegrees New Student Checklist.');
+    } else if (contentKey == "uiuc_student") {
+      return Localization().getStringEx('widget.checklist.uiuc.logged_out', 'You need to be logged in to access New Student Checklist.');
+    }
+  }
 
   @override
   State<StatefulWidget> createState() => _CheckListPanelState();